refactor(di): move container package to di and improve documentation
