+++ pkg/expressgo/di/container_test.go
@@ -1,4 +1,4 @@
-package container
+package di
 
 import (
 	"fmt"
