+++ pkg/expressgo/di/scope.go
@@ -1,4 +1,4 @@
-package container
+package di
 
 import (
 	"context"
